bitoguard_mock_api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9
httpx==0.28.1
pytest==8.4.1
PyYAML==6.0.2
